Calibrate Accelerometer Data using GGIR

Calibrate Accelerometer Data using GGIR

estimate_calibration_values(
  file,
  verbose = TRUE,
  fix_zeros = TRUE,
  fill_in = TRUE,
  by_second = FALSE,
  trim = FALSE,
  ...
)

calibrate(
  file,
  verbose = TRUE,
  fix_zeros = TRUE,
  fill_in = TRUE,
  by_second = FALSE,
  trim = FALSE,
  round_after_calibration = TRUE,
  ...
)

Arguments

file
Either a GT3X file, `AccData` object, or `data.frame` with `X/Y/Z` and `time`

verbose
print diagnostic messages, higher number result in higher verbosity

fix_zeros
Should fix_zeros be run before calculating the measures?

fill_in
if fix_zeros = TRUE, should the zeros be filled in with the last

by_second
Should the last observation carried forward be done only within the same second?

trim
if fix_zeros = TRUE, should the time course be trimmed for zero values at the beginning and the end of the time course? observation carried forward?

...
Additional arguments to pass to g.calibrate

round_after_calibration
Should the data be rounded after calibration? Will round to 3 digits

Value

A set of calibration coefficients

Examples

# \donttest{
files = read.gt3x::gt3x_datapath(2)

#> Downloading gt3x sample data from https://github.com/THLfi/read.gt3x/releases/download/v1.0/EE_left_29.5.2017-05-30.gt3x.zip

path = files[1]
res = read_actigraphy(path)

#> Error in if (tmp1[length(tmp1)] == "v" | tmp1[length(tmp1)] == "v.gz") {    dformat = 2    testcsv = read.csv(datafile, nrow = 10, skip = 10)    testcsvtopline = read.csv(datafile, nrow = 2, skip = 1)    if (ncol(testcsv) == 2 & ncol(testcsvtopline) < 4) {        mon = 2    }    else if (ncol(testcsv) >= 3 & ncol(testcsvtopline) < 4) {        mon = 3    }    else if (ncol(testcsv) >= 4 & ncol(testcsvtopline) >= 4) {        mon = 4    }} else if (tmp2[length(tmp2)] == "in") {    dformat = 1} else if (tmp3[length(tmp3)] == "av") {    dformat = 3    mon = 4} else if (tmp5[length(tmp5)] == "a") {    dformat = 4    mon = 4}: missing value where TRUE/FALSE needed

cab = calibrate(res)

#> Error in calibrate(res): object 'res' not found

# }