Run GGIR Processing

Run GGIR Processing

ggir_process(df, unit = "1 min", calibrate = TRUE, verbose = TRUE, ...)

Arguments

df
An object with columns `X`, `Y`, and `Z` or an object of class `AccData`

unit
length of time to calculate measures over. a character string specifying a time unit or a multiple of a unit to be rounded to. Valid base units are `second`, `minute`, `hour`, `day`, `week`, `month`, `bimonth`, `quarter`, `season`, `halfyear`, and `year`. Arbitrary unique English abbreviations as in the period constructor are allowed.

calibrate
Run calibration, passed to do.calibrate in GGIR

verbose
print diagnostic messages

...
additional arguments to pass to g.shell.GGIR

Value

A list with all the output elements

Examples

if (FALSE) {
url = paste0("https://github.com/THLfi/read.gt3x/files/",
"3522749/GT3X%2B.01.day.gt3x.zip")
destfile = tempfile(fileext = ".zip")
dl = utils::download.file(url, destfile = destfile, mode = "wb")
gt3x_file = utils::unzip(destfile, exdir = tempdir())
gt3x_file = gt3x_file[!grepl("__MACOSX", gt3x_file)]
path = gt3x_file
res = read_actigraphy(path)
res = fix_zeros(res)
out = ggir_process(res)
}