Calculate Summary Measures from Raw Accelerometer Data

Calculate Summary Measures from Raw Accelerometer Data

calculate_measures(
  df,
  epoch = "1 min",
  fix_zeros = TRUE,
  fill_in = TRUE,
  trim = FALSE,
  dynamic_range = c(-6, 6),
  calculate_mims = TRUE,
  verbose = TRUE,
  ...
)

calculate_ai(df, epoch = "1 min")

calculate_n_idle(df, epoch = "1 min")

calculate_mad(df, epoch = "1 min")

calculate_mims(df, epoch = "1 min", dynamic_range = c(-6, 6), ...)

Arguments

df
An object with columns `X`, `Y`, and `Z` or an object of class `AccData`

epoch
length of time to calculate measures over. a character string specifying a time unit or a multiple of a unit to be rounded to. Valid base units are `second`, `minute`, `hour`, `day`, `week`, `month`, `bimonth`, `quarter`, `season`, `halfyear`, and `year`. Arbitrary unique English abbreviations as in the period constructor are allowed.

fix_zeros
Should fix_zeros be run before calculating the measures?

fill_in
if fix_zeros = TRUE, should the zeros be filled in with the last

trim
if fix_zeros = TRUE, should the time course be trimmed for zero values at the beginning and the end of the time course? observation carried forward?

dynamic_range
Dynamic range of the device, in gravity units

calculate_mims
Should MIMS units be calculated?

verbose
print diagnostic messages

...
additional arguments to pass to [MIMSunit::mims_unit]

Value

A data set with the calculated features