Calibrate Accelerometer Data using GGIR

Calibrate Accelerometer Data using GGIR

estimate_calibration_values(
  file,
  verbose = TRUE,
  fix_zeros = TRUE,
  fill_in = TRUE,
  by_second = FALSE,
  trim = FALSE,
  ...
)

calibrate(
  file,
  verbose = TRUE,
  fix_zeros = TRUE,
  fill_in = TRUE,
  by_second = FALSE,
  trim = FALSE,
  round_after_calibration = TRUE,
  ...
)

Arguments

file
Either a GT3X file, `AccData` object, or `data.frame` with `X/Y/Z` and `time`

verbose
print diagnostic messages, higher number result in higher verbosity

fix_zeros
Should fix_zeros be run before calculating the measures?

fill_in
if fix_zeros = TRUE, should the zeros be filled in with the last

by_second
Should the last observation carried forward be done only within the same second?

trim
if fix_zeros = TRUE, should the time course be trimmed for zero values at the beginning and the end of the time course? observation carried forward?

...
Additional arguments to pass to g.calibrate

round_after_calibration
Should the data be rounded after calibration? Will round to 3 digits

Value

A set of calibration coefficients

Examples

# \donttest{
url = paste0("https://github.com/THLfi/read.gt3x/files/",
"3522749/GT3X%2B.01.day.gt3x.zip")
destfile = tempfile(fileext = ".zip")
dl = utils::download.file(url, destfile = destfile, mode = "wb")
gt3x_file = utils::unzip(destfile, exdir = tempdir())
gt3x_file = gt3x_file[!grepl("__MACOSX", gt3x_file)]
path = gt3x_file
res = read_actigraphy(path)

#> Input is a .gt3x file, unzipping to a temporary location first...

#> Unzipping gt3x data to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmpb3csO4

#> 1/1 

#> Unzipping /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmpb3csO4/GT3X+ (01 day).gt3x

#>  === info.txt, activity.bin, lux.bin extracted to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmpb3csO4/GT3X+(01day)

#> GT3X information
#>  $ Serial Number     :"NEO1DXXXXXXXX"
#>  $ Device Type       :"GT3XPlus"
#>  $ Firmware          :"2.5.0"
#>  $ Battery Voltage   :"4.22"
#>  $ Sample Rate       :30
#>  $ Start Date        : POSIXct, format: "2012-06-27 10:54:00"
#>  $ Stop Date         : POSIXct, format: "2012-06-28 11:54:00"
#>  $ Download Date     : POSIXct, format: "2012-06-28 16:25:52"
#>  $ Board Revision    :"4"
#>  $ Unexpected Resets :"0"
#>  $ Sex               :"Male"
#>  $ Height            :"172.72"
#>  $ Mass              :"69.8532249799612"
#>  $ Age               :"43"
#>  $ Race              :"White / Caucasian"
#>  $ Limb              :"Ankle"
#>  $ Side              :"Left"
#>  $ Dominance         :"Non-Dominant"
#>  $ DateOfBirth       :"621132192000000000"
#>  $ Subject Name      :"GT3XPlus"
#>  $ Serial Prefix     :"NEO"
#>  $ Last Sample Time  : 'POSIXct' num(0) 
#>  - attr(*, "tzone")= chr "GMT"
#>  $ Acceleration Scale:341
#>  $ Acceleration Min  :"-6.0"
#>  $ Acceleration Max  :"6.0"

#> Parsing GT3X data via CPP.. expected sample size: 2700000

#> Using NHANES-GT3X format - older format

#> Sample size: 2700000
#> Scaling...

#> Done (in 2.60686802864075 seconds)

cab = calibrate(res)

#> Fixing Zeros

#> Writing out file for GGIR

#> Running g.inspectfile

#> Running g.calibrate

#> 
#> Loading chunk: 1 2 3 4
#> Summary of autocalibration procedure:
#> 
#> Status: recalibration not done because not enough points on all sides of the sphere
#> Calibration error (g) before: 0.01462
#> Callibration error (g) after: 
#> Offset correction x: 0 
#> Offset correction y: 0 
#> Offset correction z: 0
#> Scale correction x: 1 
#> Scale correction y: 1 
#> Scale correction z: 1
#> Number of hours used: 24
#> Number of 10 second windows around the sphere: 5836
#> Temperature used (if available): FALSE
#> Temperature offset (if temperature is available) x:  
#> Temperature offset (if temperature is available) y:  
#> Temperature offset (if temperature is available) z: 

# }